+++ Food.h
@@ -5,7 +5,7 @@ class Food
 {
     public:
     int foodsCount = 0;
-    void addFood(World &world, RendererSFML &renderer, uint32_t x, uint32_t y);
+    void addFood(World &world, RendererSFML &renderer, uint16_t x, uint16_t y);
     void createFood(World &world, RendererSFML &renderer);
     void foodRespawn(World &world, RendererSFML &renderer);
 };
